Tape Equilibrium:
Solution using dynamic programming.
partialOperations hold:
the partial sum from 0 to item at row 0
the partial sum from max to item at row 1, this is made substracting to the accumulative sum the item[i][0] value.
the difference between row 0 and row 1 at row 2. This is what we are looking for, can be removed but was left to validate our solution.