Missing Integer:
This is a suboptimal solution based on a TreeSet.
We iterate A and add each  positive value to a TreeSet O(nlogn).
Then we iterate our TreeSet, that it's sorted to find the min integer in the array.

A solution with a hash array could be better in certain cases than this one but the spatial complexity increase.